requirements.txt
# Core LangChain (actually used)
langchain
langchain-openai
langchain-community  
langchain-core
llama-index
llama-index-core
llama-index-readers-file

# LLM Provider
openai

# Code Analysis 
tree-sitter
tree-sitter-python
GitPython

# Text Processing & Embeddings (actually used)
sentence-transformers

# Vector Store (actually used) 
faiss-cpu

# Data Processing (actually used)
numpy
pydantic
scikit-learn

# Web requests (for GitHub cloning)
requests

# Utilities (actually used)
tqdm
python-dotenv

# UI Framework
gradio

# Evaluation Metrics (actually used)
bert-score
rouge-score
nltk
ragas